requirements.txt
torch>=1.13.0
faiss-cpu>=1.7.2
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
datasets>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
